Implemented PID output filtering
The moving average implementation has been used to fully implement
filtering of the three PID controller outputs. To be able to perform
initialization of the moving averages for each PID controller, a new
automaticControl_init function has been implemented within the
automaticControl.c file. Initialization of the automaticControl is
now done in the StartControlTask function within the controlThread.c
file.